Moved server code into src/lambda
