+++ src/lambda/restaurantShow.js
@@ -1,7 +1,26 @@
 const { MongoClient, ObjectId } = require('mongodb');
-const { errorResponse, successResponse } = require('./helpers');
 const DB_URL = process.env.DB;
 const DB_NAME = process.env.DB_NAME;
+
+// // // //
+
+function errorResponse(callback, err) {
+  console.error(err);
+
+  callback(null, {
+    statusCode: 500,
+    body: JSON.stringify({ error: err })
+  })
+}
+
+function successResponse(callback, res) {
+  console.log('Saved new page request. Current count:', res.value.requests);
+
+  callback(null, {
+    statusCode: 200,
+    body: JSON.stringify(res)
+  });
+}
 
 // // // //
 
@@ -13,17 +32,13 @@ exports.handler = function(event, context, callback) {
     const db = connection.db(DB_NAME);
     const restaurantCollection = db.collection('restaurants');
 
-    console.log(event.pathParameters)
-    console.log(event.pathParameters.restaurantId)
-
     // Isolates restaurantID parameter
     // TODO - handle missing restaurantId parameter
-    const restaurantId = event.pathParameters.restaurantId
+    const restaurantId = event.queryStringParameters.id
 
     // Finds ONE restaurant
     restaurantCollection.findOne({ _id: ObjectId(restaurantId) }, (err, result) => {
       if (err) return errorResponse(callback, err);
-
 
       console.log('FOUND FOUND???')
       console.log(result)
