+++ src/lambda/violationList.js
@@ -1,22 +1,42 @@
 const { MongoClient } = require('mongodb');
-const { errorResponse, successResponse } = require('./helpers');
 const DB_URL = process.env.DB;
 const DB_NAME = process.env.DB_NAME;
 
 // // // //
 
-// GET /api/restaurants
+function errorResponse(callback, err) {
+  console.error(err);
+
+  callback(null, {
+    statusCode: 500,
+    body: JSON.stringify({ error: err })
+  })
+}
+
+function successResponse(callback, res) {
+  console.log('Saved new page request. Current count:', res.value.requests);
+
+  callback(null, {
+    statusCode: 200,
+    body: JSON.stringify(res)
+  });
+}
+
+// // // //
+
+// GET /api/violations
 exports.handler = function(event, context, callback) {
   MongoClient.connect(DB_URL, (err, connection) => {
     if (err) return errorResponse(callback, err);
 
     const db = connection.db(DB_NAME);
-    const restaurantCollection = db.collection('restaurants');
+    const violationCollection = db.collection('violations');
 
     // GET /api/cities
-    restaurantCollection.distinct('address.city', (err, result) => {
+    violationCollection.find({}, (err, result) => {
         if (err) return errorResponse(callback, err);
 
+        // TODO - this result should be cached
         // res.setHeader('Cache-Control', 'max-age=604800, public');
         callback(null, {
           statusCode: 200,
@@ -26,5 +46,4 @@ exports.handler = function(event, context, callback) {
     })
 
   });
-
 }
